@@ -49,34 +49,34 @@
   },
   "homepage": "https://github.com/getsaf/shallow-render#readme",
   "peerDependencies": {
-    "@angular/common": "15.x",
-    "@angular/compiler": "15.x",
-    "@angular/core": "15.x",
-    "@angular/forms": "15.x",
-    "@angular/platform-browser": "15.x"
+    "@angular/common": "16.x",
+    "@angular/compiler": "16.x",
+    "@angular/core": "16.x",
+    "@angular/forms": "16.x",
+    "@angular/platform-browser": "16.x"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.4",
-    "@angular-eslint/eslint-plugin": "^14.0.2",
-    "@angular-eslint/eslint-plugin-template": "^14.0.2",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/compiler-cli": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@types/jest": "^28.1.6",
-    "@typescript-eslint/eslint-plugin": "^5.8.0",
-    "@typescript-eslint/parser": "^5.8.0",
-    "eslint": "^8.5.0",
-    "jest": "^28.1.3",
-    "jest-preset-angular": "^12.2.0",
-    "prettier": "^2.0.5",
-    "rxjs": "~7.5.0",
-    "ts-node": "^10.4.0",
-    "typescript": "~4.8.2",
-    "zone.js": "~0.12.0"
+    "@angular-devkit/build-angular": "^16.1.5",
+    "@angular-eslint/eslint-plugin": "^16.1.0",
+    "@angular-eslint/eslint-plugin-template": "^16.1.0",
+    "@angular/common": "^16.1.0",
+    "@angular/compiler": "^16.1.0",
+    "@angular/compiler-cli": "^16.1.0",
+    "@angular/core": "^16.1.0",
+    "@angular/forms": "^16.1.0",
+    "@angular/platform-browser": "^16.1.0",
+    "@angular/platform-browser-dynamic": "^16.1.0",
+    "@angular/router": "^16.1.0",
+    "@types/jest": "^29.5.3",
+    "@typescript-eslint/eslint-plugin": "^6.1.0",
+    "@typescript-eslint/parser": "^6.1.0",
+    "eslint": "^8.45.0",
+    "jest": "^29.6.1",
+    "jest-preset-angular": "^13.1.1",
+    "prettier": "^3.0.0",
+    "rxjs": "~7.8.0",
+    "ts-node": "^10.9.1",
+    "typescript": "~5.1.3",
+    "zone.js": "~0.13.0"
   }
 }
